mcp-client/main.py
import uvicorn
from app import app
from config import settings
from dotenv import load_dotenv

load_dotenv()

if __name__ == "__main__":
    uvicorn.run(
        app,
        host=settings.api_host,
        port=settings.api_port
    )