mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "MCP Client API Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "openai>=2.16.0",
    "python-dotenv>=1.1.0",
    "fastapi>=0.115.12",
    "pydantic>=2.11.2",
    "pydantic-settings>=2.8.1",
    "uvicorn>=0.34.0",
]
